+++ crates/torii/mcp/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+edition.workspace = true
+name = "torii-mcp"
+version.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+sqlx.workspace = true
+tokio.workspace = true
+torii-sqlite.workspace = true
