+++ crates/torii/mcp/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod types;
+pub mod tools;
+pub mod resources;
