+++ crates/torii/server/Cargo.toml
@@ -6,6 +6,8 @@ version.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+torii-mcp = { path = "../../torii/mcp" }
+
 anyhow.workspace = true
 base64.workspace = true
 camino.workspace = true
