+++ crates/torii/server/src/handlers/sql.rs
@@ -1,12 +1,11 @@
 use std::net::IpAddr;
 use std::sync::Arc;
 
-use base64::engine::general_purpose::STANDARD;
-use base64::Engine;
 use http::header::CONTENT_TYPE;
 use hyper::{Body, Method, Request, Response, StatusCode};
 use include_str;
-use sqlx::{Column, Row, SqlitePool, TypeInfo};
+use sqlx::SqlitePool;
+use torii_sqlite::utils::map_row_to_json;
 
 use super::Handler;
 
@@ -116,45 +115,4 @@ impl Handler for SqlHandler {
     async fn handle(&self, req: Request<Body>, _client_addr: IpAddr) -> Response<Body> {
         self.handle_request(req).await
     }
-}
-
-pub fn map_row_to_json(row: &sqlx::sqlite::SqliteRow) -> serde_json::Value {
-    let mut obj = serde_json::Map::new();
-    for (i, column) in row.columns().iter().enumerate() {
-        let value: serde_json::Value = match column.type_info().name() {
-            "TEXT" => row
-                .get::<Option<String>, _>(i)
-                .map_or(serde_json::Value::Null, serde_json::Value::String),
-            "INTEGER" => row
-                .get::<Option<i64>, _>(i)
-                .map_or(serde_json::Value::Null, |n| serde_json::Value::Number(n.into())),
-            "REAL" => row.get::<Option<f64>, _>(i).map_or(serde_json::Value::Null, |f| {
-                serde_json::Number::from_f64(f)
-                    .map_or(serde_json::Value::Null, serde_json::Value::Number)
-            }),
-            "BLOB" => row.get::<Option<Vec<u8>>, _>(i).map_or(serde_json::Value::Null, |bytes| {
-                serde_json::Value::String(STANDARD.encode(bytes))
-            }),
-            _ => {
-                // Try different types in order
-                if let Ok(val) = row.try_get::<i64, _>(i) {
-                    serde_json::Value::Number(val.into())
-                } else if let Ok(val) = row.try_get::<f64, _>(i) {
-                    serde_json::json!(val)
-                } else if let Ok(val) = row.try_get::<bool, _>(i) {
-                    serde_json::Value::Bool(val)
-                } else if let Ok(val) = row.try_get::<String, _>(i) {
-                    serde_json::Value::String(val)
-                } else {
-                    // Handle or fallback to BLOB as base64
-                    let val = row.get::<Option<Vec<u8>>, _>(i);
-                    val.map_or(serde_json::Value::Null, |bytes| {
-                        serde_json::Value::String(STANDARD.encode(bytes))
-                    })
-                }
-            }
-        };
-        obj.insert(column.name().to_string(), value);
-    }
-    serde_json::Value::Object(obj)
 }
